+++ satellite_determination/event_finder/event_finder_rhodesmill/event_finder_rhodesmill.py
@@ -11,7 +11,7 @@ from satellite_determination.custom_dataclasses.position_time import PositionTim
 from satellite_determination.custom_dataclasses.overhead_window import OverheadWindow
 from satellite_determination.custom_dataclasses.reservation import Reservation
 from satellite_determination.custom_dataclasses.time_window import TimeWindow
-from satellite_determination.event_finder.event_finder_rhodesmill.overhead_window_slew import AntennaPosition, \
+from satellite_determination.event_finder.event_finder_rhodesmill.support.overhead_window_slew import AntennaPosition, \
     OverheadWindowSlew
 from satellite_determination.event_finder.event_finder_rhodesmill.support.overhead_window_from_events import \
     EventRhodesmill, EventTypesRhodesmill, OverheadWindowFromEvents
