+++ tests/event_finder/event_finder_rhodesmill/test_overhead_window_slew_unsorted_antenna_positions.py
@@ -3,7 +3,7 @@ from typing import List
 
 from satellite_determination.custom_dataclasses.position_time import PositionTime
 from satellite_determination.custom_dataclasses.time_window import TimeWindow
-from satellite_determination.event_finder.event_finder_rhodesmill.overhead_window_slew import AntennaPosition, \
+from satellite_determination.event_finder.event_finder_rhodesmill.support.overhead_window_slew import AntennaPosition, \
     OverheadWindowSlew
 from tests.event_finder.event_finder_rhodesmill.definitions import ARBITRARY_ANTENNA_POSITION, ARBITRARY_FACILITY
